requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.2
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cmake==3.26.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
einops==0.6.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.3
flatbuffers==23.3.3
flit_core==3.8.0
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gast==0.4.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.13.3
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
jax==0.4.6
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter-events==0.6.3
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.1
libclang==16.0.0
lightning-utilities==0.8.0
lit==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
pkgutil_resolve_name==1.3.10
platformdirs==3.1.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.0
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qudida==0.0.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.1
scipy==1.9.1
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==67.5.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.2.28
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.2
torch==2.0.0
torchaudio==0.7.0a0+ac17b64
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.3
triton==2.0.0
typeguard==3.0.2
typing_extensions==4.4.0
urllib3==1.26.15
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.14.1
yarl==1.8.2
zipp==3.15.0
